BackUp_project(2)/Project/back_end/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
Pillow
jinja2
aiofiles